Reannounce to tracker before removing torrent
Been using the plugin and noticed that sometimes the tracker isn't fully updated before auto-removing a torrent (e.g. uploaded 500MB but tracker says 450MB).
So I added a functionality to pause and resume a torrent before removing it so the tracker will get updated stats.
Thought about 'component.get("Core").force_reannounce(torrent_id)', but it doesn't work ok (Next announce is not updated so I think it's somehow broken).
The sleep delay is there to give the tracker a chance to be updated, though I think it can be optimized.